refactor: delete preview page
