+++ src/components/index.ts
@@ -1,23 +1,25 @@
 import NotFound from '../pages/notFound/NotFound.tsx';
-import ButtonLink from './buttonLink/ButtonLink.tsx';
 import Header from './header/Header.tsx';
 import Layout from './layout/Layout.tsx';
 import LoginRegisterLayout from './layout/login-register-layout.tsx';
 import Loading from './loading/Loading.tsx';
 import AuthInput from './login_register_inputs/auth_input.tsx';
 import FormInput from './login_register_inputs/form_input.tsx';
+import Logo from './logo/Logo.tsx';
 import Navigation from './navigation/Navigation.tsx';
 import Saver from './saver/saver.tsx';
+import Title from './title/Title.tsx';
 
 export {
   AuthInput,
-  ButtonLink,
   FormInput,
   Header,
   Layout,
   Loading,
   LoginRegisterLayout,
+  Logo,
   Navigation,
   NotFound,
   Saver,
+  Title,
 };
