+++ src/utils/CONSTANTS.ts
@@ -1,7 +1,6 @@
 const CONSTANTS = {
   home: '/',
   profile: '/profile',
-  preview: '/preview',
   shop: '/shop',
   login: '/login',
   register: '/register',
